+++ packages/watcher/src/receiver/EventReceiver.ts
@@ -1,4 +1,4 @@
-import { NativeKEvent } from '../types/types';
+import { NativeKEvent } from '../types/NativeKEvent';
 import { AxiosInstance } from 'axios';
 import { JsonStreamParser } from '../json-parser/JsonStreamParser';
 import { EventDispatcher } from '../dispatcher/EventDispatcher';
